Move from removed ProcessBuilder to Process
The main change in this commit is the replacement of
Symfony\Component\Process\ProcessBuilder by the direct
use of  Symfony\Component\Process\Process instead.

The Symfony\Component\Process\ProcessBuilder class was
deprecated in Symfony 3.4 and removed in 4.0.

This is a huge change because we were relying on the, now
removed class, near everywhere.

At the same time, adjust all tests to work with PHPUnit 9,
add param and return types and other small details.